Add fractional team support and lookback auto-detection
- Support fractional team members (0.5 FTE) in --team-change flag
- Add intelligent lookback period auto-detection with 'auto' default
- Improve scenario descriptions to show capacity impact percentages
- Fix Sprint Health scope change calculation (was showing 700%+)
- Add comprehensive test coverage for new features
- Document heuristics in HEURISTICS.md
- Update CLI to support both 'auto' and numeric lookback values

+++ src/infrastructure/velocity_adjustment_parser.py
@@ -101,11 +101,11 @@ class VelocityAdjustmentParser:
                 change_str = parts["change"]
                 # Handle explicit + or - prefix
                 if change_str.startswith("+"):
-                    change = int(change_str[1:])
+                    change = float(change_str[1:])
                 elif change_str.startswith("-"):
-                    change = -int(change_str[1:])
+                    change = -float(change_str[1:])
                 else:
-                    change = int(change_str)
+                    change = float(change_str)
 
                 if change == 0:
                     raise ValueError("Change must be non-zero")
